front: fix display of players.
Before there was nothing printed due to 2 bugs.

First, players in game compnonent was a type of Object instead of Array
I've fixed it with data.players.

The second bug was a type bug. PlayerId was a string and player.id was a
number so we had to cast PlayerId in number